v0.8.0 | Squashed commit of the following:
commit 6ed028e8f8c059fb649eb3c32095b61248344b12
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Sun Oct 20 20:55:41 2024 +0200

    Fix type instability in 'Annotations'.

commit cbe65358230ad71419f65b1ee150a3289502c5c6
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Sun Oct 20 19:20:32 2024 +0200

    Fix bug in 'FeedForwardNetwork'.

commit b4b347e2842b3dc75b8213bb55ddb06b663eb60e
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Sun Oct 20 19:01:51 2024 +0200

    Add support for example weights in 'Trainer' and 'SequenceClassifierTransformer'.

commit 0450562502d384f794a1869f0429918913f15d01
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Sun Oct 20 17:22:15 2024 +0200

    Implement 'Dataset.add_balancing_weights()' method.

commit 134222ef88244739245f81bf24bfefc1a7aa736b
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Sun Oct 20 14:53:06 2024 +0200

    Refactor 'newsuse.ml.models'.

commit fdae8dc0b3bf8bbcdec48c32e360fa409038cb52
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Sun Oct 20 03:55:37 2024 +0200

    Implement flexible iterative annotations framework.

commit d1d6c0cba98fd8d19235e3d2fc34580900505382
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Sat Oct 19 20:34:19 2024 +0200

    Implement 'newsuse.ml.models.classifiers' module with 'AutoModelForSequenceClassification', 'NewsuseSequenceClassifier' and 'SetFitModel'.

commit 609193aa2dda5f585357f283ff03bd469e5961da
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Tue Oct 15 16:09:38 2024 +0200

    Various updates.

commit 25f251eca06c410c506b7185a738dd9a91406b3d
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Tue Oct 15 14:11:42 2024 +0200

    Implement consistent training API using 'newsuse.ml.models.sequenceclassification.py' and 'newsuse.ml.models.setfit'.

commit 7962a425635253e2df1e38b8d8f1cfa420422ca2
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Tue Oct 15 09:21:46 2024 +0200

    Rename 'Dataset.concatenate' => 'Dataset.concat'.

commit 4da910bc2cc20ac818b8f19edfe49525f75e1ca4
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Tue Oct 15 09:12:55 2024 +0200

    Fix bug that 'Trainer' was deleting checkpoint folders even when starting training from a checkpoint.

commit c3bea024f1e45862f9d2fe88469c4596e167067b
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Tue Oct 15 09:12:20 2024 +0200

    Implement 'Dataset.make_splits'.

commit a5e6808f2a451f811f7a011e0f30d15b2c74a051
Author: Szymon Talaga <stalaga@protonmail.com>
Date:   Tue Oct 15 09:11:45 2024 +0200

    Implement annotator selection in 'Annotations'.

+++ newsuse/ml/models/trainer.py
@@ -1,7 +1,8 @@
+from abc import ABC, abstractmethod
 from collections.abc import Callable, Iterable
 from pathlib import Path
 from shutil import rmtree
-from typing import Any
+from typing import Any, Self
 
 import evaluate
 import numpy as np
@@ -10,6 +11,12 @@ from transformers import EvalPrediction
 from transformers.trainer import TrainOutput
 
 from newsuse.types import PathLike
+
+from ..datasets import Dataset, DatasetDict
+
+
+class TrainingArguments(transformers.TrainingArguments):
+    pass
 
 
 class Trainer(transformers.Trainer):
@@ -51,9 +58,11 @@ class Trainer(transformers.Trainer):
 
         return compute_metrics
 
-    def train(self, *args: Any, **kwargs: Any) -> TrainOutput:
+    def train(
+        self, resume_from_checkpoint: str | bool | None = None, *args: Any, **kwargs: Any
+    ) -> TrainOutput:
         outdir = Path(self.args.output_dir)
-        if outdir.exists():
+        if not resume_from_checkpoint and outdir.exists():
             for checkpoint in outdir.glob("checkpoint-*"):
                 if checkpoint.is_file():
                     checkpoint.unlink()
@@ -71,3 +80,53 @@ class Trainer(transformers.Trainer):
                     else:
                         rmtree(path)
         super().save_model(output_dir, *args, **kwargs)
+
+    def _set_signature_columns_if_needed(self) -> None:
+        super()._set_signature_columns_if_needed()
+        if "weight" in self.train_dataset.column_names:
+            self._signature_columns = [*self._signature_columns, "weight"]
+
+
+class TrainingMixin(ABC):
+    @classmethod
+    @abstractmethod
+    def factory(cls, *args: Any, **kwargs: Any) -> Callable[..., Self]:
+        """Get ``model_init()`` factory function."""
+
+    @classmethod
+    def get_tokenizer(
+        cls, name_or_path: str, *args: Any, **kwargs: Any
+    ) -> transformers.PreTrainedTokenizerBase:
+        """Get model tokenizer."""
+        return transformers.AutoTokenizer.from_pretrained(name_or_path, *args, **kwargs)
+
+    @classmethod
+    def get_training_arguments(cls, *args: Any, **kwargs: Any) -> TrainingArguments:
+        """Get :class:`transformers.TrainingArguments` instance."""
+        return TrainingArguments(*args, **kwargs)
+
+    @classmethod
+    def get_trainer(
+        cls,
+        *,
+        args: TrainingArguments,
+        model: Self | None = None,
+        model_init: Callable[..., Self] | None = None,
+        tokenizer: transformers.PreTrainedTokenizerBase | None = None,
+        **kwargs: Any,
+    ) -> Trainer:
+        """Get :class:`Trainer` instance."""
+        return Trainer(
+            args=args, model=model, model_init=model_init, tokenizer=tokenizer, **kwargs
+        )
+
+    @classmethod
+    def preprocess_dataset(
+        cls,
+        dataset: Dataset | DatasetDict,
+        tokenizer: transformers.PreTrainedTokenizerBase,
+        *args: Any,  # noqa
+        **kwargs: Any,  # noqa
+    ) -> Dataset:
+        """Preprocess ``dataset`` before training."""
+        return dataset.tokenize(tokenizer)
